Rename and restructure ptsa.data.readers
Updating to use PEP8 naming conventions and organizing things that
logically fit together into single modules instead of one class
per module as before.
