+++ ptsa/data/filters/MonopolarToBipolarMapper.py
@@ -119,7 +119,7 @@ def main_fcn():
     # selecting only one session
     base_events = base_events[base_events.eegfile == base_events[0].eegfile]
 
-    from ptsa.data.readers.TalReader import TalReader
+    from ptsa.data.readers.tal import TalReader
 
     tal_path = '/Users/m/data/eeg/R1060M/tal/R1060M_talLocs_database_bipol.mat'
     tal_reader = TalReader(filename=tal_path)
