+++ ptsa/data/filters/MorletWaveletFilter.py
@@ -1,5 +1,3 @@
-__author__ = 'm'
-
 import time
 
 import numpy as np
@@ -12,7 +10,7 @@ from ptsa.data.filters import BaseFilter
 from ptsa.wavelet import morlet_multi, next_pow2
 
 
-class MorletWaveletFilter(PropertiedObject,BaseFilter):
+class MorletWaveletFilter(PropertiedObject, BaseFilter):
     """
     Applies a Morlet wavelet transform to a time series, returning the power and phase spectra over time.
 
@@ -34,8 +32,6 @@ class MorletWaveletFilter(PropertiedObject,BaseFilter):
     verbose: bool
         Print out the wavelet parameters
     """
-
-
     _descriptors = [
         TypeValTuple('freqs', np.ndarray, np.array([], dtype=np.float)),
         TypeValTuple('width', int, 5),
@@ -255,330 +251,3 @@ class MorletWaveletFilter(PropertiedObject,BaseFilter):
             print('total time wavelet loop: ', time.time() - wavelet_start)
 
         return self.build_output_arrays(wavelet_pow_array, wavelet_phase_array, time_axis)
-
-
-def test_1_old():
-    import time
-    start = time.time()
-
-    e_path = '/Users/m/data/events/RAM_FR1/R1060M_events.mat'
-
-    from ptsa.data.readers import BaseEventReader
-
-    base_e_reader = BaseEventReader(event_file=e_path, eliminate_events_with_no_eeg=True, use_ptsa_events_class=False)
-
-    base_e_reader.read()
-
-    base_events = base_e_reader.get_output()
-
-    base_events = base_events[base_events.type == 'WORD']
-
-    # selecting only one session
-    base_events = base_events[base_events.eegfile == base_events[0].eegfile]
-
-    from ptsa.data.readers.TalReader import TalReader
-    tal_path = '/Users/m/data/eeg/R1060M/tal/R1060M_talLocs_database_bipol.mat'
-    tal_reader = TalReader(tal_filename=tal_path)
-    monopolar_channels = tal_reader.get_monopolar_channels()
-    bipolar_pairs = tal_reader.get_bipolar_pairs()
-
-    print('bipolar_pairs=', bipolar_pairs)
-
-    from ptsa.data.experimental.TimeSeriesSessionEEGReader import TimeSeriesSessionEEGReader
-
-    # time_series_reader = TimeSeriesSessionEEGReader(events=base_events, channels = ['002', '003', '004', '005'])
-    time_series_reader = TimeSeriesSessionEEGReader(events=base_events, channels=monopolar_channels)
-    ts_dict = time_series_reader.read()
-
-    first_session_data = list(ts_dict.items())[0][1]
-
-    print(first_session_data)
-
-    wavelet_start = time.time()
-
-    wf = MorletWaveletFilter(time_series=first_session_data,
-                             freqs=np.logspace(np.log10(3), np.log10(180), 2),
-                             # freqs=np.array([3.]),
-                             output='power',
-                             # resamplerate=50.0
-                             )
-
-    pow_wavelet, phase_wavelet = wf.filter()
-    print('wavelet total time = ', time.time() - wavelet_start)
-    # return pow_wavelet
-
-    from ptsa.data.experimental.EventDataChopper import EventDataChopper
-    edcw = EventDataChopper(events=base_events, event_duration=1.6, buffer=1.0,
-                            data_dict={base_events[0].eegfile: pow_wavelet})
-
-    chopped_wavelets = edcw.filter()
-
-    chopped_wavelets = list(chopped_wavelets.items())[0][1]  # getting first item of return dictionary
-
-    print('total time = ', time.time() - start)
-    #
-    # from ptsa.data.filters.ResampleFilter import ResampleFilter
-    # rsf = ResampleFilter (resamplerate=50.0)
-    # rsf.set_input(chopped_wavelets)
-    # chopped_wavelets_resampled = rsf.filter()
-    #
-    # return chopped_wavelets_resampled
-    return chopped_wavelets
-
-
-def test_2_old():
-    import time
-    start = time.time()
-
-    e_path = '/Users/m/data/events/RAM_FR1/R1060M_events.mat'
-
-    from ptsa.data.readers import BaseEventReader
-
-    base_e_reader = BaseEventReader(event_file=e_path, eliminate_events_with_no_eeg=True, use_ptsa_events_class=False)
-
-    base_e_reader.read()
-
-    base_events = base_e_reader.get_output()
-
-    base_events = base_events[base_events.type == 'WORD']
-
-    # selecting only one session
-    base_events = base_events[base_events.eegfile == base_events[0].eegfile]
-
-    from ptsa.data.readers.TalReader import TalReader
-    tal_path = '/Users/m/data/eeg/R1060M/tal/R1060M_talLocs_database_bipol.mat'
-    tal_reader = TalReader(tal_filename=tal_path)
-    monopolar_channels = tal_reader.get_monopolar_channels()
-    bipolar_pairs = tal_reader.get_bipolar_pairs()
-
-    print('bipolar_pairs=', bipolar_pairs)
-
-    from ptsa.data.experimental.TimeSeriesEEGReader import TimeSeriesEEGReader
-
-    time_series_reader = TimeSeriesEEGReader(events=base_events, start_time=0.0,
-                                             end_time=1.6, buffer_time=1.0, keep_buffer=True)
-
-    base_eegs = time_series_reader.read(channels=monopolar_channels)
-
-    # base_eegs = base_eegs[:, 0:10, :]
-    # bipolar_pairs = bipolar_pairs[0:10]
-
-
-    wf = MorletWaveletFilter(time_series=base_eegs,
-                             freqs=np.logspace(np.log10(3), np.log10(180), 2),
-                             # freqs=np.array([3.]),
-                             output='power',
-                             # resamplerate=50.0
-                             )
-
-    pow_wavelet, phase_wavelet = wf.filter()
-
-    print('total time = ', time.time() - start)
-
-    res_start = time.time()
-
-    # from ptsa.data.filters.ResampleFilter import ResampleFilter
-    # rsf = ResampleFilter (resamplerate=50.0)
-    # rsf.set_input(pow_wavelet)
-    # pow_wavelet = rsf.filter()
-
-
-
-    print('resample_time=', time.time() - res_start)
-    return pow_wavelet
-
-
-def test_2():
-    import time
-    start = time.time()
-
-    e_path = '/Users/m/data/events/RAM_FR1/R1060M_events.mat'
-
-    from ptsa.data.readers import BaseEventReader
-
-    base_e_reader = BaseEventReader(filename=e_path)
-
-    base_events = base_e_reader.read()
-
-    base_events = base_events[base_events.type == 'WORD']
-
-    # selecting only one session
-    base_events = base_events[base_events.eegfile == base_events[0].eegfile]
-
-    from ptsa.data.readers.TalReader import TalReader
-    tal_path = '/Users/m/data/eeg/R1060M/tal/R1060M_talLocs_database_bipol.mat'
-    tal_reader = TalReader(filename=tal_path)
-    monopolar_channels = tal_reader.get_monopolar_channels()
-    bipolar_pairs = tal_reader.get_bipolar_pairs()
-
-    print('bipolar_pairs=', bipolar_pairs)
-
-    from ptsa.data.experimental.TimeSeriesEEGReader import TimeSeriesEEGReader
-
-    time_series_reader = TimeSeriesEEGReader(events=base_events, start_time=0.0,
-                                             end_time=1.6, buffer_time=1.0, keep_buffer=True)
-
-    base_eegs = time_series_reader.read(channels=monopolar_channels)
-
-    # base_eegs = base_eegs[:, 0:10, :]
-    # bipolar_pairs = bipolar_pairs[0:10]
-
-
-    wf = MorletWaveletFilter(time_series=base_eegs,
-                             freqs=np.logspace(np.log10(3), np.log10(180), 2),
-                             # freqs=np.array([3.]),
-                             output='power',
-                             )
-
-    pow_wavelet, phase_wavelet = wf.filter()
-
-    print('total time = ', time.time() - start)
-
-    res_start = time.time()
-
-    # from ptsa.data.filters.ResampleFilter import ResampleFilter
-    # rsf = ResampleFilter (resamplerate=50.0)
-    # rsf.set_input(pow_wavelet)
-    # pow_wavelet = rsf.filter()
-
-
-
-    print('resample_time=', time.time() - res_start)
-    return pow_wavelet
-
-
-def test_1():
-    import time
-    start = time.time()
-
-    e_path = '/Users/m/data/events/RAM_FR1/R1060M_events.mat'
-
-    from ptsa.data.readers import BaseEventReader
-
-    base_e_reader = BaseEventReader(filename=e_path)
-
-    base_events = base_e_reader.read()
-
-    base_events = base_events[base_events.type == 'WORD']
-
-    # selecting only one session
-    base_events = base_events[base_events.eegfile == base_events[0].eegfile]
-
-    from ptsa.data.readers.TalReader import TalReader
-    tal_path = '/Users/m/data/eeg/R1060M/tal/R1060M_talLocs_database_bipol.mat'
-    tal_reader = TalReader(filename=tal_path)
-    monopolar_channels = tal_reader.get_monopolar_channels()
-    bipolar_pairs = tal_reader.get_bipolar_pairs()
-
-    dataroot = base_events[0].eegfile
-    from ptsa.data.readers import EEGReader
-    session_reader = EEGReader(session_dataroot=dataroot, channels=monopolar_channels)
-    session_eegs = session_reader.read()
-
-    wavelet_start = time.time()
-
-    wf = MorletWaveletFilter(time_series=session_eegs,
-                             freqs=np.logspace(np.log10(3), np.log10(180), 2),
-                             # freqs=np.array([3.]),
-                             output='power',
-                             # resamplerate=50.0
-                             )
-
-    pow_wavelet, phase_wavelet = wf.filter()
-    print('wavelet total time = ', time.time() - wavelet_start)
-    # return pow_wavelet
-
-
-    from ptsa.data.filters import EventDataChopper
-    sedc = EventDataChopper(events=base_events, session_data=pow_wavelet, start_time=0.0, end_time=1.6, buffer_time=1.0)
-    chopped_wavelets = sedc.filter()
-
-    print('total time = ', time.time() - start)
-
-    return chopped_wavelets
-
-
-if __name__ == '__main__':
-    edcw_1 = test_1()
-    edcw_2 = test_2()
-
-    wavelet_1 = edcw_1[0, 0, 0, 500:1300]
-    wavelet_2 = edcw_2[0, 0, 0, 500:1300]
-
-    import matplotlib;
-
-    matplotlib.use('Qt4Agg')
-
-    import matplotlib.pyplot as plt
-
-    plt.get_current_fig_manager().window.raise_()
-
-    plt.plot(np.arange(wavelet_1.shape[0]) - 1, wavelet_1, 'k')
-    plt.plot(np.arange(wavelet_2.shape[0]) - 1, wavelet_2, 'r--')
-
-    plt.show()
-    #
-
-# if __name__ == '__main__':
-#     edcw_1 = test_1()
-#     edcw_2 = test_2()
-#
-#     wavelet_1 = edcw_1[0,0,0,500:1300]
-#     wavelet_2 = edcw_2[0,0,0,500:1300]
-#
-#     import matplotlib;
-#     matplotlib.use('Qt4Agg')
-#
-#
-#     import matplotlib.pyplot as plt
-#     plt.get_current_fig_manager().window.raise_()
-#
-#
-#     plt.plot(np.arange(wavelet_1.shape[0])-1,wavelet_1,'r--')
-#     plt.plot(np.arange(wavelet_2.shape[0]),wavelet_2)
-#
-#
-#     plt.show()
-
-
-#
-#     print
-
-# if __name__=='__main__':
-#
-#
-#
-#     edcw = test_1()
-#     pow_wavelet = test_2()
-#
-#     import matplotlib;
-#     matplotlib.use('Qt4Agg')
-#
-#
-#     import matplotlib.pyplot as plt
-#     plt.get_current_fig_manager().window.raise_()
-#
-#
-#
-#     # pow_wavelet_res =  resample(pow_wavelet.data[0,1,3,500:-501],num=180)[50:]
-#     # edcw_res =  resample(edcw.data[0,1,3,500:-501],num=180)[50:]
-#
-#     # pow_wavelet_res =  resample(pow_wavelet.data[0,1,3,:],num=180)[50:-50]
-#     # edcw_res =  resample(edcw.data[0,1,3,:],num=180)[50:-50]
-#
-#     pow_wavelet_res =  pow_wavelet.data[0,1,3,50:-50]
-#     edcw_res =  edcw.data[0,1,3,50:-50]
-#
-#
-#
-#     print
-#
-#
-#
-#
-#     plt.plot(np.arange(pow_wavelet_res.shape[0]),pow_wavelet_res)
-#     plt.plot(np.arange(edcw_res.shape[0])-0.5,edcw_res)
-#
-#
-#     plt.show()
-#
