--- ptsa/data/readers/BaseReader.py
@@ -1,5 +1,0 @@
-__author__ = 'm'
-
-class BaseReader(object):
-    def read(self):
-        pass
