+++ ptsa/data/readers/__init__.py
@@ -1,11 +1,9 @@
-from .BaseReader import BaseReader
-from .BaseEventReader import BaseEventReader
-from .EEGReader import EEGReader
-from .NetCDF4XrayReader import NetCDF4XrayReader
-from .PTSAEventReader import PTSAEventReader
-from .ParamsReader import ParamsReader
-from .TalReader import TalReader
-from .TalStimOnlyReader import TalStimOnlyReader
-from .BaseRawReader import BaseRawReader
-from .CMLEventReader import CMLEventReader
-from .IndexReader import JsonIndexReader
+from .base import *
+from .eeg import EEGReader
+from .events import *
+from .hdf5 import H5RawReader
+from .index import JsonIndexReader
+from .netcdf import NetCDF4XrayReader
+from .params import ParamsReader
+from .raw import BaseRawReader
+from .tal import *
