+++ ptsa/data/readers/eeg.py
@@ -5,44 +5,49 @@ import warnings
 import numpy as np
 import xarray as xr
 
-from ptsa.data.common import TypeValTuple, PropertiedObject
-from ptsa.data.readers.ParamsReader import ParamsReader
-from ptsa.data.readers.BaseRawReader import BaseRawReader
-from ptsa.data.readers.H5RawReader import H5RawReader
-from ptsa.data.readers import BaseReader
 from ptsa import six
+from ptsa.data.common import TypeValTuple, PropertiedObject
+from ptsa.data.readers.params import ParamsReader
+from ptsa.data.readers.raw import BaseRawReader
+from ptsa.data.readers.hdf5 import H5RawReader
+from ptsa.data.readers import BaseReader
 from ptsa.data.TimeSeriesX import TimeSeriesX
 
+__all__ = [
+    'EEGReader',
+]
 
+
+# FIXME: centralize PTSA exception classes
 class IncompatibleDataError(Exception):
     pass
 
 
 class EEGReader(PropertiedObject, BaseReader):
-    """
-    Reader that knows how to read binary eeg files. It can read chunks of the eeg signal based on events input
-    or can read entire session if session_dataroot is non empty.
+    """Reader that knows how to read binary eeg files. It can read chunks of the
+    eeg signal based on events input or can read entire session if
+    session_dataroot is non empty.
 
     Keyword Arguments
     -----------------
     channels : np.ndarray
         numpy array of channel labels
     start_time : float
-        read start offset in seconds w.r.t to the eegeffset specified in the events recarray
+        read start offset in seconds w.r.t to the eegeffset specified in the
+        events recarray
     end_time:
-        read end offset in seconds w.r.t to the eegeffset specified in the events recarray
+        read end offset in seconds w.r.t to the eegeffset specified in the
+        events recarray
     buffer_time : float
-        extra buffer in seconds (subtracted from start read and added to end read)
+        extra buffer in seconds (subtracted from start read and added to end
+        read)
     events : np.recarray
         numpy recarray representing Events
     session_dataroot : str
-        path to session dataroot. When set the reader will read the entire session
+        path to session dataroot. When set the reader will read the entire
+        session
     remove_bad_events : bool
         Remove "bad" events. Defaults to True.
-
-    Returns
-    -------
-    None
 
     """
     _descriptors = [
