+++ ptsa/data/readers/hdf5.py
@@ -1,7 +1,13 @@
-from ptsa.data.readers.BaseRawReader import BaseRawReader
-import numpy as np
-import h5py
 import os.path as osp
+
+import h5py
+import numpy as np
+
+from ptsa.data.readers.raw import BaseRawReader
+
+__all__ = [
+    'H5RawReader',
+]
 
 
 class H5RawReader(BaseRawReader):
@@ -121,10 +127,3 @@ class H5RawReader(BaseRawReader):
                 raise RuntimeError("All eventdata is nan!")
 
             return eventdata, read_ok_mask
-
-
-if __name__ == "__main__":
-    filename = osp.expanduser("~/mnt/rhino/data/eeg/R1275D/behavioral/FR1/session_0/host_pc/20170531_170954/eeg_timeseries.h5")
-    channels = np.array(['%.03d' % i for i in range(1, 10)])
-    with h5py.File(filename, 'r') as hfile:
-        reader = H5RawReader.read_h5file(hfile, channels, [0], 1000)
