+++ tests/test_classifier.py
@@ -5,7 +5,7 @@ import pytest
 
 from ptsa.data.readers import BaseEventReader
 from ptsa.data.filters.MorletWaveletFilter import MorletWaveletFilter
-from ptsa.data.readers.TalReader import TalReader
+from ptsa.data.readers.tal import TalReader
 from ptsa.data.readers import EEGReader
 from ptsa.data.filters import MonopolarToBipolarMapper
 from ptsa.test.utils import get_rhino_root, skip_without_rhino
